mok: move the mok_state_variables definitions to their own header
This lets us access the definitions for this structure, and the data
being used at runtime, from unit tests.
